refactor: Improve controller performance
This refactors the controller code to improve the performance and
optimize the number of calls made to the API server.

In the original implementation it created several goroutines per
`StackSet`. Each goroutine would poll the same resources resulting in a
lot of unessesary Get calls to the API server. The number of calls
increased with the number of `StackSets` managed by the controller.
The following calls were made:

    n x [Get Endpoints]
    n x [Get Ingress] x 3
    n x [List Stacks] x 5
    n x [Get Deployment]
    n x [Get Service]
    n x [Get HPA]

Where `n` is the number of `StackSets`. This sums up to `12 x n` `Get`
calls for each iteration of the control loops.

The new implementation fetches all the resources in the main loop and
organizes them in `StackSetContainers` which is a struct storing all the
resources related to a single StackSet. The `StackContainers` are then
passed to individual reconciler loops which works the same as before
except the read from the in memory representation instead of fetching
reasources from the API server. This reduces the number of `Get` calls
from `12 x n` to `6` for each iteration of the control loop.

Signed-off-by: Mikkel Oscar Lyderik Larsen <mikkel.larsen@zalando.de>