Use equality.Semantic.DeepEqual for resource comparison
Use `equality.Semantic.DeepEqual` instead of `reflect.DeepEqual` for
correct comparison of special Kubernetes types like `resource.Quantity`.

The `resource.Quantity` type can represent the same value in different
representations i.e. `2000m` is equal to `2`. In that case the internal
fields of the struct will look differently and therefore it's not
possible to compare just with `reflect.DeepEqual`.

Signed-off-by: Mikkel Oscar Lyderik Larsen <mikkel.larsen@zalando.de>