Use resource generation to determine updates
Before we tried to determine if a child resource, like a deployment of a
stack, needed updates by comparing the original resource with the
resource after we made potental changes.
This doesn't work in most cases because when a resource is submitted it
potentially goes through admission controllers which changes the
defaults thus modifying the resource. Especially for deployments created
for a stack this didn't work and the stackset-controller is constantly
updating deployment resources even though it's not needed and not actual
changes are applied.

Instead use the Generation on the resources to determine if the child
resource needs an update or not. This is done by storing an annotation
with the "parent" resource generation and then comparing this value when
reconciling a resource.

Fix #49

Signed-off-by: Mikkel Oscar Lyderik Larsen <mikkel.larsen@zalando.de>