Use averageValue for skipper based autoscaler
This sets the field `averageValue` for the skipper based autoscaler.

This requires
https://github.com/zalando-incubator/kube-metrics-adapter/releases/tag/v0.0.4
which adds support for `averageValue` such that average is calculated by
the HPA rather than by the kube-metrics-adapter. This makes the
resulting average more reliable for scaling.

Note this means the stackset-controller only supports Kubernetes
`>=v1.14.0` if the skipper autoscaler is used.

Signed-off-by: Mikkel Oscar Lyderik Larsen <mikkel.larsen@zalando.de>