Change references from Nginx to Skipper
The e2e tests and documentation files are using Nginx docker image as an
example. It makes e2e test to pull multiple times this image from
Dockerhub. We are reaching their [rate-limit][0] and our e2e tests are
breaking frequently.

This commit changes all the references to the Nginx image to the skipper
image, including documentation.

[0]: https://docs.docker.com/docker-hub/download-rate-limit/