Stack's RouteGroups inherit annotations (#297)
* Stack's RouteGroups inherit annotations

The RouteGroups generated per Stack are not receiving the annotations
from StackSet's RouteGroups definition. It can lead to a wide range of
bug reports and is a non-expected behaviour.

This commit adds the procedure to copy the annotations from the
StackSet's RouteGroups spec and tests.

* Add stack annotations e2e testing

The E2E tests aren't checking whether annotations are being inherit from
StackSets' Ingress and RouteGroups Spec to the Stack's resulting Ingress
and RouteGroups.

This commit adds the E2E tests that verify this aspect as well as fix
the `run_e2e.sh` script to use the `--cluster-domain` flags. This script
is only used locally.