Allow traffic switch before stack is 100% Ready (#456)
Given the need for switching traffic to a stack that isn't 100% ready
yet, the minReadyPercent StackSet field allows setting a minimum
percentage of ready replicas required to enable the traffic switch.

minReadyPercent is optional and expects values from 1 to 100. In case of
value equal to or under 0 or over 100, it's normalized to 100, which
keeps the original behaviour.