Add the type match label (#365)
When using the `autoscaler` we are still letting `kube-metrics-adapter`
to rely in the metric name. This commit adds the `type` label to the
ZMON and SQS queue len metrics. It also sorts the metrics before
generating them, based on different criteria.

Previously, sort was based just on the HPA metric type. Now sort is done
by `StackSet`'s `autoscaler` metric type and once it conflicts
`Average`, `AverageUtilization`, etc. It guarantees we support multiple
metrics of the same type and do not regenerate the HPA if Kubernetes
sends the metrics in a different order.

Signed-off-by: Jonathan Juares Beber <jonathanbeber@gmail.com>
Signed-off-by: Mikkel Oscar Lyderik Larsen <mikkel.larsen@zalando.de>
Co-authored-by: Katyanna Moura <amelie.kn@gmail.com>