Support dedicated ConfigMap created by other means
To give the user the possibility of creating the configMap to be attached to the
Stack by themselves, Stackset Controller checks if there's a reference in the
ConfigurationResources before creating the versioned ConfigMap.
If there isn't, the existing resource named on ConfigurationResources.Name is
updated with the Stack info on ownerReferences.