Reference ConfigMap for per stack versioning (#520)
Currently Stacks share the same configuration resources, which may lead to confusion and incidents. As a first step on having one configuration resource dedicated to each stack, this Pull Request implements ConfigMaps versioned by reference. Other resources will follow in future Pull Requests.

The expected behaviour is that given the a ConfigMap named in the <stack-name>-<configmap-name> pattern is referenced under ConfigurationResources on the Stack definition, Stackset Controller identifies the deployed ConfigMap and updates it with the Stack info on ownerReferences.